+++ app/models/marty/data_grid.rb
@@ -105,7 +105,7 @@ class Marty::DataGrid < Marty::Base
     end
   end
 
-  has_mcfly
+  mcfly
 
   validates :name, :data, :metadata, presence: true